more code review comments addressed

+++ engine-plugins/graph-plugins/src/main/scala/org/trustedanalytics/atk/plugins/orientdb/OrientDbTypeConverter.scala
@@ -47,9 +47,8 @@ object OrientDbTypeConverter {
   def convertOrientDbtoDataType(orientDbType: OType): DataType = orientDbType match {
     case OType.LONG => DataTypes.toDataType("int64")
     case OType.INTEGER => DataTypes.toDataType("int32")
-    case OType.FLOAT => DataTypes.toDataType("float32")
+    case OType.FLOAT => DataTypes.toDataType("float64")
     case OType.STRING => DataTypes.toDataType("string")
     case _ => throw new IllegalArgumentException(s"Unable to convert $orientDbType to DataType")
   }
-
 }
